client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "invoke>=2.2.0",
    "pyinstaller>=6.14.1",
    "python-dotenv>=1.1.0",
    "pywebview>=5.4",
]
